backend/requirements_replit.txt
fastapi==0.112.0
uvicorn[standard]==0.35.0
python-dotenv==1.1.1
python-multipart==0.0.9
pydantic==2.7.1
websockets==12.0
anthropic==0.18.1
aiofiles==23.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4 